MD3 Implementation Progress
Added ability to issue drift warning based on change in margin density
relative to that of reference distribution.
Next step is to implement system to collect labeled samples to confirm
that drift is occurring (or rule out drift).
Another next step is to work on building out example(s) of MD3, with SVM
specifically for now.